requirement.txt
astroid==1.6.6
backports.functools-lru-cache==1.5
configparser==3.8.1
Django==1.8.4
django-cors-headers==2.4.1
djangorestframework==3.6.4
enum34==1.1.6
futures==3.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
psycopg2-binary==2.8.3
pylint==1.9.5
pylint-django==0.11.1
pylint-plugin-utils==0.5
pytz==2019.2
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.11.2
